Рефакторинг..
