+++ js/effects-helper.js
