+++ js/form-validation-helper.js
@@ -1,6 +1,7 @@
-import { isEscapeKey } from './full-size-evt-helper.js';
-import { setImgScale, imgContainer } from './img-scale-evt-helper.js';
-import { hideSlider, setVisibleImageStyle } from './nouislider-evt-helper.js';
+import { previewShowImg } from './full-size-evt-helper.js';
+import { setImgScale } from './img-scale-evt-helper.js';
+import { hideSlider, setVisibleImageStyle } from './effects-helper.js';
+import { isEscapeKey } from './util.js';
 const uploadFileControl = document.querySelector('#upload-file');
 const uploadOverlay = document.querySelector('.img-upload__overlay');
 const imgUploadCancel = document.querySelector('.img-upload__cancel');
@@ -10,7 +11,6 @@ const commentInput = document.querySelector('.text__description');
 const MAX_COMMENT_LENGTH = 140;
 const TAG_REGEX = /^#[а-яёa-z0-9]{1,19}$/i;
 const MAX_TAGS_LENGTH = 5;
-const FILE_TYPES = ['jpg', 'jpeg', 'png'];
 
 const pristine = new Pristine(form,
   {
@@ -117,13 +117,7 @@ const setValidationEvt = (submit) => {
   setSubmitListener(submit);
   uploadFileControl.addEventListener('change', () => {
 
-    const file = uploadFileControl.files[0];
-    const fileName = file.name.toLowerCase();
-    const matches = FILE_TYPES.some((it) => fileName.endsWith(it));
-    if (matches) {
-      imgContainer.querySelector('img').src = URL.createObjectURL(file);
-    }
-
+    previewShowImg();
     setImgScale();
     hideSlider();
     setVisibleImageStyle();
