+++ js/util.js
@@ -1,8 +1,3 @@
-//функция рандома в указанном диапазоне
-export const getRandomInteger = (min, max) => {
-  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
-  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
-  const result = Math.random() * (upper - lower + 1) + lower;
+const isEscapeKey = (evt) => evt.key === 'Escape';
 
-  return Math.floor(result);
-};
+export {isEscapeKey};
